issue-148 Clean up files with "2" in them
resolves #148
