+++ web/src/components/EditUserCell/EditUserCell.mock.js
@@ -1,6 +1,6 @@
 // Define your own mock data here:
 export const standard = (/* vars, { ctx, req } */) => ({
-  part2: {
+  editUser: {
     id: 42,
   },
 })
