+++ web/src/components/PartCell/PartCell.mock.js
@@ -1,6 +1,6 @@
 // Define your own mock data here:
 export const standard = (/* vars, { ctx, req } */) => ({
-  editUser2: {
+  part: {
     id: 42,
   },
 })
