+++ web/src/layouts/MainLayout/MainLayout.js
@@ -157,7 +157,7 @@ const MainLayout = ({ children }) => {
               {isAuthenticated && data?.user?.userName ? (
                 <Link
                   className="h-full w-full"
-                  to={routes.newPart2({ userName: data?.user?.userName })}
+                  to={routes.newPart({ userName: data?.user?.userName })}
                 >
                   <Svg name="plus" className="text-indigo-300 w-full h-full" />
                 </Link>
@@ -210,14 +210,14 @@ const MainLayout = ({ children }) => {
               }}
             >
               <div className="p-4 w-40">
-                <Link to={routes.user2({ userName: data?.user?.userName })}>
+                <Link to={routes.user({ userName: data?.user?.userName })}>
                   <h3 className="text-indigo-800" style={{ fontWeight: '500' }}>
                     Hello {data?.user?.name}
                   </h3>
                 </Link>
                 <hr />
                 <br />
-                <Link to={routes.editUser2({ userName: data?.user?.userName })}>
+                <Link to={routes.editUser({ userName: data?.user?.userName })}>
                   <div className="text-indigo-800">Edit Profile</div>
                 </Link>
                 <a href="#" className="text-indigo-800" onClick={logOut}>
