--- web/src/pages/EditPart2Page/EditPart2Page.stories.js
@@ -1,7 +1,0 @@
-import EditPart2Page from './EditPart2Page'
-
-export const generated = () => {
-  return <EditPart2Page />
-}
-
-export default { title: 'Pages/EditPart2Page' }
