+++ web/src/pages/EditPartPage/EditPartPage.stories.js
@@ -1,0 +1,7 @@
+import EditPartPage from './EditPartPage'
+
+export const generated = () => {
+  return <EditPartPage />
+}
+
+export default { title: 'Pages/EditPartPage' }
