--- web/src/pages/NewPart2Page/NewPart2Page.stories.js
@@ -1,7 +1,0 @@
-import NewPart2Page from './NewPart2Page'
-
-export const generated = () => {
-  return <NewPart2Page />
-}
-
-export default { title: 'Pages/NewPart2Page' }
