+++ web/src/pages/NewPartPage/NewPartPage.stories.js
@@ -1,0 +1,7 @@
+import NewPartPage from './NewPartPage'
+
+export const generated = () => {
+  return <NewPartPage />
+}
+
+export default { title: 'Pages/NewPartPage' }
