--- web/src/pages/Part2Page/Part2Page.stories.js
@@ -1,7 +1,0 @@
-import Part2Page from './Part2Page'
-
-export const generated = () => {
-  return <Part2Page />
-}
-
-export default { title: 'Pages/Part2Page' }
