+++ web/src/pages/PartPage/PartPage.stories.js
@@ -1,0 +1,7 @@
+import PartPage from './PartPage'
+
+export const generated = () => {
+  return <PartPage />
+}
+
+export default { title: 'Pages/PartPage' }
